requirements.txt
torch
torchvision
torchaudio
opencv-python
Pillow
numpy
pandas
seaborn
PyYAML
tqdm
ultralytics